Moved fns to common
