+++ test/unit/include/test_unit/printing.hpp
@@ -13,19 +13,9 @@
 namespace boost {
 namespace mysql {
 
-// character set
-struct character_set;
-bool operator==(const character_set& lhs, const character_set& rhs);
-std::ostream& operator<<(std::ostream& os, const character_set& v);
-
 // address_type
 enum class address_type;
 std::ostream& operator<<(std::ostream& os, address_type value);
-
-// errcode_with_diagnostics
-struct errcode_with_diagnostics;
-bool operator==(const errcode_with_diagnostics& lhs, const errcode_with_diagnostics& rhs);
-std::ostream& operator<<(std::ostream& os, const errcode_with_diagnostics& v);
 
 namespace detail {
